+++ src/frontend/admin/src/translations/pages/orders/breadcrumbsTranslations.ts
@@ -16,4 +16,9 @@ export const ordersBreadcrumbsTranslation = defineMessages({
     defaultMessage: "View",
     description: "Breadcrumb label for the order view page",
   },
+  createBreadcrumb: {
+    id: "pages.admin.orders.breadcrumbsTranslations.createBreadcrumb",
+    defaultMessage: "Create",
+    description: "Breadcrumb label for the order create page",
+  },
 });