+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to
 
 ### Added
 
+- Create standalone to_own orders in admin
 - Handle deep link in back office
 - Add client API endpoint to return a deep link value in `OfferingViewSet`
 - Add admin API to manage offering deep links for organizations